moved NeighbourhoodBlock unit test class to separate test module
